+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupRepository.java
@@ -2,7 +2,5 @@ package kitchenpos.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.MenuGroup;
-
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
 }
