+++ src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.menugroup.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuGroupResponse {
