+++ src/test/java/kitchenpos/menugroup/domain/MenuGroupTest.java
@@ -5,8 +5,6 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import kitchenpos.domain.MenuGroup;
-
 @DisplayName("메뉴 그룹 도메인 테스트")
 public class MenuGroupTest {
